Refactor explore code
